requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
PyPDF2
python-docx
transformers
torch
numpy
requests
beautifulsoup4
lxml